+++ Client/src/app/pages/layout/components/home/home.component.less
@@ -1,30 +1,91 @@
-@import "@taiga-ui/core/styles/taiga-ui-local.less";
 
-:host {
-  display: block;
+#home_container{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
 }
 
-.stepper {
-  margin-bottom: 2rem;
+#myCourse{
+    width: 100%;
+    padding: 30px 30px 0 30px;
+    
 }
 
-.uppercase-name {
-  text-transform: uppercase;
+#myProfie{
+    background-color: green;
+    border: 1px solid black;
+    width: 350px;
+    height: 100%;
 }
 
-.account {
-  display: flex;
-  flex: 1;
-  justify-content: space-between;
+#header{
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    .title{
+        display: flex;
+        flex-direction: column;
+    }
+    .search{
+        width: 35%;
+    }
+    h1,p{
+        margin: 0;
+    }
+    p{
+        margin-top: 7px;
+    }
 }
 
-.ticks-labels {
-  .tui-slider-ticks-labels();
 
-  color: var(--tui-text-02);
-  margin-top: 0.25rem;
-}
 
-.tui-container {
-  padding: 0.625rem;
+
+
+#body{
+    margin-top: 20px;
+    width: 100%;
+    .onGoing{
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        .tui-space_bottom-3{
+            margin: 0;
+        }
+    }
+
+    .box{
+        width: 60rem;
+        height: 39rem;
+        margin-top: 20px;
+    }
+
+    .list_course{
+        padding: 0 10px;
+        display: flex;
+        flex-direction: row;
+        justify-content: space-evenly;
+        flex-wrap: wrap;
+        ::ng-deep .t-top, ::ng-deep .t-actions{
+            height: 0 ;
+        }
+        ::ng-deep .t-heading{
+            margin: 0;
+        }
+
+        ::ng-deep .t-content{
+            margin-top: 5px;
+        }
+
+        .cart_course{
+            margin-bottom: 20px;
+        }
+
+        .btn_course{
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+            margin: 5px 0;
+        }
+    }
 }
