+++ Client/src/app/pages/layout/components/profile/profile.component.less
@@ -1,0 +1,30 @@
+@import "@taiga-ui/core/styles/taiga-ui-local.less";
+
+:host {
+  display: block;
+}
+
+.stepper {
+  margin-bottom: 2rem;
+}
+
+.uppercase-name {
+  text-transform: uppercase;
+}
+
+.account {
+  display: flex;
+  flex: 1;
+  justify-content: space-between;
+}
+
+.ticks-labels {
+  .tui-slider-ticks-labels();
+
+  color: var(--tui-text-02);
+  margin-top: 0.25rem;
+}
+
+.tui-container {
+  padding: 0.625rem;
+}
